@@ -78,16 +78,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.1",
-    "@angular/cdk": "^7.1.1",
-    "@angular/common": "^7.1.1",
-    "@angular/compiler": "^7.1.1",
-    "@angular/core": "^7.1.1",
-    "@angular/forms": "^7.1.1",
-    "@angular/http": "^7.1.1",
-    "@angular/platform-browser": "^7.1.1",
-    "@angular/platform-browser-dynamic": "^7.1.1",
-    "@angular/router": "^7.1.1",
+    "@angular/animations": "^8.2.10",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.10",
+    "@angular/compiler": "^8.2.10",
+    "@angular/core": "^8.2.10",
+    "@angular/forms": "^8.2.10",
+    "@angular/http": "^7.2.15",
+    "@angular/platform-browser": "^8.2.10",
+    "@angular/platform-browser-dynamic": "^8.2.10",
+    "@angular/router": "^8.2.10",
     "@fortawesome/fontawesome-svg-core": "^1.2.8",
     "@fortawesome/free-solid-svg-icons": "^5.5.0",
     "core-js": "^2.5.7",
@@ -97,52 +97,52 @@
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.11.1",
-    "@angular-devkit/build-ng-packagr": "^0.11.1",
-    "@angular/cli": "^7.1.1",
-    "@angular/compiler-cli": "^7.1.1",
-    "@angular/language-service": "^7.1.1",
+    "@angular-devkit/build-angular": "^0.803.9",
+    "@angular-devkit/build-ng-packagr": "^0.803.9",
+    "@angular/cli": "^8.2.10",
+    "@angular/compiler-cli": "^8.2.10",
+    "@angular/language-service": "^8.2.10",
     "@compodoc/compodoc": "^1.1.7",
     "@types/hammerjs": "^2.0.36",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "^2.0.6",
     "@types/mousetrap": "^1.6.0",
-    "@types/node": "^10.12.12",
+    "@types/node": "^12.7.12",
     "codeclimate-test-reporter": "^0.5.0",
-    "codelyzer": "~4.5.0",
-    "coveralls": "^3.0.1",
+    "codelyzer": "~5.1.2",
+    "coveralls": "^3.0.7",
     "cpr": "^3.0.1",
-    "cross-env": "^5.2.0",
-    "husky": "^1.2.0",
-    "jasmine-core": "~2.99.1",
+    "cross-env": "^6.0.3",
+    "husky": "^3.0.9",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.1.3",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage": "^1.1.2",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage": "^2.0.1",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-edge-launcher": "^0.4.2",
-    "karma-firefox-launcher": "^1.1.0",
+    "karma-firefox-launcher": "^1.2.0",
     "karma-ie-launcher": "^1.0.0",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "karma-mocha-reporter": "^2.2.5",
     "karma-safari-launcher": "^1.0.0",
-    "karma-sonarqube-unit-reporter": "0.0.18",
-    "lite-server": "^2.3.0",
-    "ng-packagr": "^4.4.5",
+    "karma-sonarqube-unit-reporter": "0.0.21",
+    "lite-server": "^2.5.4",
+    "ng-packagr": "^5.6.1",
     "prettier": "=1.14.3",
     "pretty-quick": "=1.8.0",
-    "protractor": "~5.4.1",
-    "rimraf": "^2.6.2",
-    "sass-lint": "^1.12.1",
-    "snyk": "^1.114.0",
-    "sonar-scanner": "^3.0.3",
-    "ts-node": "~7.0.0",
-    "tsickle": "^0.34.0",
-    "tslib": "^1.9.0",
-    "tslint": "~5.11.0",
-    "typedoc": "^0.13.0-0",
-    "typescript": "~3.1.1"
+    "protractor": "~5.4.2",
+    "rimraf": "^3.0.0",
+    "sass-lint": "^1.13.1",
+    "snyk": "^1.234.2",
+    "sonar-scanner": "^3.1.0",
+    "ts-node": "~8.4.1",
+    "tsickle": "^0.37.0",
+    "tslib": "^1.10.0",
+    "tslint": "~5.20.0",
+    "typedoc": "^0.15.0",
+    "typescript": ">=3.4.0 <3.6.0"
   }
 }
